requirements.txt
requests
asgiref==3.2.3
Django==3.0.3
Pillow==7.0.0
python-dotenv==0.11.0
pytz==2019.3
sqlparse==0.3.0
google-auth==1.11.2
requests==2.22.0